Added AaveShorter interface and contract
Contract takes given parameters and opens a short using Aave and
Uniswap. The short is constructed by borrowing the intial amount passed
times the short factor using a flash loan. The borrowed assets are then
swapped for more of the given colateral. The output of the swap and the
given colateral are then deposited into aave as colateral for the flash
loan that becomes a debt token.